wallet-ext: fix importing webextension polyfill in dapp interface
