+++ apps/wallet/src/shared/api-env.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+export enum API_ENV {
+    local = 'local',
+    devNet = 'devNet',
+    testNet = 'testNet',
+    customRPC = 'customRPC',
+}
