+++ apps/wallet/src/ui/app/components/account-address/index.tsx
@@ -3,11 +3,11 @@
 
 import cl from 'classnames';
 
-import { API_ENV } from '_app/ApiProvider';
 import CopyToClipboard from '_components/copy-to-clipboard';
 import ExplorerLink from '_components/explorer-link';
 import { ExplorerLinkType } from '_components/explorer-link/ExplorerLinkType';
 import { useAppSelector, useMiddleEllipsis } from '_hooks';
+import { API_ENV } from '_src/shared/api-env';
 
 import st from './AccountAddress.module.scss';
