+++ apps/wallet/src/ui/app/components/explorer-link/Explorer.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { API_ENV, DEFAULT_API_ENV } from '_app/ApiProvider';
+import { DEFAULT_API_ENV } from '_app/ApiProvider';
+import { API_ENV } from '_src/shared/api-env';
 
 import type { ObjectId, SuiAddress, TransactionDigest } from '@mysten/sui.js';
