+++ apps/wallet/src/ui/app/redux/slices/app/index.ts
@@ -4,7 +4,7 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 
 import { AppType } from './AppType';
-import { DEFAULT_API_ENV, type API_ENV } from '_app/ApiProvider';
+import { DEFAULT_API_ENV } from '_app/ApiProvider';
 import {
     clearForNetworkSwitch,
     fetchAllOwnedAndRequiredObjects,
@@ -13,6 +13,7 @@ import {
 import type { PayloadAction } from '@reduxjs/toolkit';
 import type { RootState } from '_redux/RootReducer';
 import type { NetworkEnvType } from '_src/background/NetworkEnv';
+import type { API_ENV } from '_src/shared/api-env';
 import type { AppThunkConfig } from '_store/thunk-extras';
 
 type AppState = {
